+++ src/components/forms/DeleteTask.js
@@ -3,7 +3,7 @@ import { jsx } from 'theme-ui';
 import { Button } from '@theme-ui/components';
 import PropTypes from 'prop-types';
 import { useContext } from 'react';
-import TaskContext from '../context/TaskContext';
+import TaskContext from '../../context/TaskContext';
 
 const DeleteTask = ({ task }) => {
   const { deleteTask } = useContext(TaskContext);
